refactor: 불필요한 패키지 구조 제거
