+++ src/main/java/com/atwoz/member/domain/member/profile/Position.java
@@ -1,7 +1,7 @@
 package com.atwoz.member.domain.member.profile;
 
-import com.atwoz.member.exception.exceptions.member.profile.contact.LatitudeRangeException;
-import com.atwoz.member.exception.exceptions.member.profile.contact.LongitudeRangeException;
+import com.atwoz.member.exception.exceptions.member.profile.LatitudeRangeException;
+import com.atwoz.member.exception.exceptions.member.profile.LongitudeRangeException;
 import jakarta.persistence.Embeddable;
 import java.math.BigDecimal;
 import lombok.AccessLevel;
